Change Heading to Title-Case and Add Stars for Fun
There is a compelling argument that the greeting phrase in "hello world"
programs should be a fully grammatically-correct sentence, "Hello
world.", still others will argue that "world" should also be
capitalized. Is punctuation important? These are all formatting and
presentation standards problems problems that arise when constructing a
"hello world" program.

One that often goes criminally un-discussed is a question of line
breaks. Leaving aside the multitude of incompatible, and often
system-dependant methods for representing a "new line", many programmers
make the foolish mistake of not considering the possibility that their
"hello world" program should not contain any such character or sequence
in its output.

As for this project, I believe that Title-Case will improve the
professional appearance the authors are trying to convey. Furthermore,
punctuation is often inappropriate in titles that are shorter than a
full line-width, so the period shall remain absent.

I also added stars. For fun!